+++ HlslDecompiler.Tests/ShaderSources/ps_4_0/dxbc_continue.fx
@@ -1,7 +1,7 @@
 float4 main(float4 texcoord : TEXCOORD) : SV_Target
 {
 	float4 t0 = 0;
-	float t1 = 0;
+	int t1 = 0;
 	while (true) {
 		if (t1 >= 8) {
 			break;
